backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
sse-starlette==2.2.1
redis==6.4.0
celery==5.5.3
asyncpg==0.30.0
sqlalchemy==2.0.43
httpx==0.28.1
firebase-admin==6.8.0
google-cloud-firestore==2.21.0
google-cloud-aiplatform==1.112.0
google-cloud-texttospeech==2.27.0
google-cloud-storage==2.19.0
